Added typevalidation to true in sequelize constructor; Moved EntityType to ContractEntity model
