+++ test/repositories/ContractEntityRepository/save.test.ts
@@ -1,13 +1,9 @@
 import BigNumber from 'bignumber.js';
 import assert from '../../test_utils/assert';
 
-import {
-  EntityType,
-} from '../../../src/repositories/ContractEntityRepository';
-
 import Repositories from '../../../src/repositories/Repositories';
 
-import ContractEntity from '../../../src/models/ContractEntity';
+import ContractEntity, { EntityType } from '../../../src/models/ContractEntity';
 import Util from './util';
 
 import chai = require('chai');
@@ -65,17 +61,17 @@ describe('ContractEntityRepository::save', (): void => {
   });
 
   it('should fail when invalid entity type is to be saved', async (): Promise<void> => {
+    const invalidEntityType = 'invalid_entity_type' as EntityType;
     const contractEntity = new ContractEntity(
       '0x0000000000000000000000000000000000000002',
-      'invalid_entity_type' as EntityType,
+      invalidEntityType,
       new BigNumber(1),
       createdAt,
     );
 
     assert.isRejected(
       config.repos.contractEntityRepository.save(contractEntity),
-      'Invalid entity type'
+      `${invalidEntityType}`
     );
-
   });
 });
